+++ src/KegMonitor.Server/Models/HX711Payload.cs
@@ -2,8 +2,8 @@
 {
     internal class HX711Payload
     {
-        public long Weight { get; set; }
-        public long WeightRaw { get; set; }
-        public long AbsRaw { get; set; }
+        public int Weight { get; set; }
+        public int WeightRaw { get; set; }
+        public int AbsRaw { get; set; }
     }
 }
